+++ src/main/java/dev/shadowsoffire/placebo/util/SpecialTooltipItem.java
@@ -1,0 +1,10 @@
+package dev.shadowsoffire.placebo.util;
+
+/**
+ * Marker interface for items that have special tooltip interactions.
+ * <p>
+ * If the currently hovered item implements this interface, tooltip scroll wheel logic will be activated.
+ */
+public interface SpecialTooltipItem {
+
+}